+++ Common/Systems/WandSystem.cs
@@ -11,7 +11,7 @@ namespace ImproveGame.Common.Systems
 
         // 液体法杖
         public static bool AbsorptionMode = false;
-        public static byte LiquidMode = 0;
+        public static short LiquidMode = 0;
 
         // 自动钓鱼机
         public static bool SelectPoolMode = false;
